requirements.txt
nibabel>=3.2.0
numpy>=1.19.0
scipy>=1.7.0
mne>=1.0.0
matplotlib>=3.4.0
pandas>=1.3.0
# Medical image registration (ANTsPy)
antspyx>=0.4.0

